refactor(dto): move dto to proper place
