+++ packages/dto/associations/GetSubventionsResponseDto.ts
@@ -1,4 +1,4 @@
-import { DemandeSubvention } from "../search/DemandeSubvention";
+import { DemandeSubvention } from "../demandeSubvention";
 
 export interface GetSubventionsSuccessResponseDto {
     subventions: DemandeSubvention[] | null;
