+++ packages/dto/search/index.ts
@@ -1,3 +1,2 @@
 export * from "./AssociationNameDto";
-export * from "./DemandeSubvention";
 export * from "./SearchCodeError";
